Fixed Long to Integer GA

+++ server/http-api/src/main/java/app/socketiot/server/api/model/GoogleAssistant/GoogleAssistantTokenRes.java
@@ -3,10 +3,10 @@ package app.socketiot.server.api.model.GoogleAssistant;
 public class GoogleAssistantTokenRes {
     public String token_type;
     public String access_token;
-    public Long expires_in;
+    public Integer expires_in;
     public String refresh_token;
 
-    public GoogleAssistantTokenRes(String token_type, String access_token, String refresh_token, Long expires_in) {
+    public GoogleAssistantTokenRes(String token_type, String access_token, String refresh_token, Integer expires_in) {
         this.token_type = token_type;
         this.access_token = access_token;
         this.expires_in = expires_in;
